Добавлен и настроен класс PageController
